+++ nav2_ros_common/include/nav2_ros_common/node_utils.hpp
@@ -26,6 +26,7 @@
 #include "rcl_interfaces/srv/list_parameters.hpp"
 #include "pluginlib/exceptions.hpp"
 #include "ament_index_cpp/get_package_share_directory.hpp"
+#include "ament_index_cpp/get_package_share_path.hpp"
 
 #ifdef __APPLE__
   #include <pthread.h>
@@ -439,8 +440,7 @@ inline void replaceOrAddArgument(
 inline std::string get_package_share_directory(const std::string & package_name)
 {
   #if RCLCPP_VERSION_GTE(30, 1, 4)
-  std::filesystem::path pkg_share_dir;
-  ament_index_cpp::get_package_share_directory(package_name, pkg_share_dir);
+  std::filesystem::path pkg_share_dir = ament_index_cpp::get_package_share_path(package_name);
   return pkg_share_dir.string();
   #else
   return ament_index_cpp::get_package_share_directory(package_name);
