+++ nav2_smoother/test/CMakeLists.txt
@@ -1,3 +1,21 @@
+
+# Build dummy smoother plugin for testing
+add_library(dummy_smoother SHARED dummy_smoother.cpp)
+target_link_libraries(dummy_smoother
+  nav2_core::nav2_core
+  rclcpp::rclcpp
+  rclcpp_lifecycle::rclcpp_lifecycle
+  nav2_ros_common::nav2_ros_common
+  pluginlib::pluginlib
+)
+
+install(TARGETS dummy_smoother
+  EXPORT dummy_smoother
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
+)
+
 nav2_add_gtest(test_smoother_server
   test_smoother_server.cpp
 )
